move signingscript.worker functions into signingscript.task
worker.py was left over from when we initially forked signingscript from
signingworker.  The taskcluster and worker portions became scriptworker,
and the signing portions became signingscript.  After that split,
worker.py was a misnomer, and occasionally confused the author when he
was dealing with both scriptworker and signingscript edits.  Therefore,
this is a non-critical refactor that will nonetheless improve things in
the signingscript world.
