+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Changed
+- moved signingscript.worker functions into signingscript.task
 
 ## [0.4.1] - 2016-08-16
 ### Fixed
